requirements.txt
accelerate==1.6.0
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
clip==1.0
docopt==0.6.2
elastic-transport==8.17.1
elasticsearch==9.0.0
filelock==3.18.0
fsspec==2025.3.2
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.30.2
idna==3.10
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==1.36.0
networkx==3.3
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==11.2.1
pillow_heif==0.22.0
pipreqs==0.4.13
platformdirs==4.3.7
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pycountry==24.6.1
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
PyExifTool==0.5.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
reverse-geocoder==1.5.1
rpds-py==0.24.0
safetensors==0.5.3
scipy==1.15.2
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
sympy==1.13.3
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchvision==0.22.0+cu126
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
yarg==0.1.10
